--- common/src/main/kotlin/Constants.kt
@@ -1,5 +1,0 @@
-import kotlinx.datetime.Instant
-
-private val INSTANT_NONE = Instant.fromEpochMilliseconds(Long.MIN_VALUE)
-val Instant.Companion.NONE
-    get() = INSTANT_NONE
