--- common/src/main/kotlin/models/IctrlCommand.kt
@@ -1,10 +1,0 @@
-package models
-
-enum class IctrlCommand {
-    NONE,
-    CREATE,
-    READ,
-    UPDATE,
-    DELETE,
-    SEARCH,
-}
