--- common/src/main/kotlin/models/IctrlCommodityFilter.kt
@@ -1,6 +1,0 @@
-package models
-
-data class IctrlCommodityFilter(
-    var searchString: String = "",
-    var ownerId: IctrlUserId = IctrlUserId.NONE,
-)
