--- common/src/main/kotlin/models/IctrlCommodityId.kt
@@ -1,10 +1,0 @@
-package models
-
-@JvmInline
-value class IctrlCommodityId(private val id: String) {
-    fun asString() = id
-
-    companion object {
-        val NONE = IctrlCommodityId("")
-    }
-}
