--- common/src/main/kotlin/models/IctrlCommodityPermissionClient.kt
@@ -1,10 +1,0 @@
-package models
-
-enum class IctrlCommodityPermissionClient {
-    READ,
-    UPDATE,
-    DELETE,
-    MAKE_VISIBLE_PUBLIC,
-    MAKE_VISIBLE_GROUP,
-    MAKE_VISIBLE_OWNER,
-}
