--- common/src/main/kotlin/models/IctrlCommodityType.kt
@@ -1,8 +1,0 @@
-package models
-
-enum class IctrlCommodityType {
-    TUBE_LINE_PART,
-    FASTENER_PART,
-    COATING_MATERIAL,
-    NONE,
-}
