--- common/src/main/kotlin/models/IctrlError.kt
@@ -1,9 +1,0 @@
-package models
-
-data class IctrlError(
-    val code: String = "",
-    val group: String = "",
-    val field: String = "",
-    val message: String = "",
-    val exception: Throwable? = null,
-)
