--- common/src/main/kotlin/models/IctrlProductId.kt
@@ -1,10 +1,0 @@
-package models
-
-@JvmInline
-value class IctrlProductId(private val id: String) {
-    fun asString() = id
-
-    companion object {
-        val NONE = IctrlProductId("")
-    }
-}
