--- common/src/main/kotlin/models/IctrlRequestId.kt
@@ -1,10 +1,0 @@
-package models
-
-@JvmInline
-value class IctrlRequestId(private val id: String) {
-    fun asString() = id
-
-    companion object {
-        val NONE = IctrlRequestId("")
-    }
-}
