--- common/src/main/kotlin/models/IctrlState.kt
@@ -1,7 +1,0 @@
-package models
-
-enum class IctrlState {
-    NONE,
-    RUNNING,
-    FAILING,
-}
