--- common/src/main/kotlin/models/IctrlUserId.kt
@@ -1,10 +1,0 @@
-package models
-
-@JvmInline
-value class IctrlUserId(private val id: String) {
-    fun asString() = id
-
-    companion object {
-        val NONE = IctrlUserId("")
-    }
-}
