--- common/src/main/kotlin/models/IctrlVisibility.kt
@@ -1,8 +1,0 @@
-package models
-
-enum class IctrlVisibility {
-    NONE,
-    VISIBLE_TO_OWNER,
-    VISIBLE_TO_GROUP,
-    VISIBLE_PUBLIC,
-}
