--- common/src/main/kotlin/models/IctrlWorkMode.kt
@@ -1,7 +1,0 @@
-package models
-
-enum class IctrlWorkMode {
-    PROD,
-    TEST,
-    STUB,
-}
