+++ common/src/main/kotlin/ru/otus/otuskotlin/common/Constants.kt
@@ -1,0 +1,7 @@
+package ru.otus.otuskotlin.common
+
+import kotlinx.datetime.Instant
+
+private val INSTANT_NONE = Instant.fromEpochMilliseconds(Long.MIN_VALUE)
+val Instant.Companion.NONE
+    get() = INSTANT_NONE
