+++ common/src/main/kotlin/ru/otus/otuskotlin/common/IctrlContext.kt
@@ -1,0 +1,21 @@
+package ru.otus.otuskotlin.common
+
+import kotlinx.datetime.Instant
+import ru.otus.otuskotlin.common.models.*
+import ru.otus.otuskotlin.common.stubs.IctrlStubs
+
+data class IctrlContext(
+    var command: IctrlCommand = IctrlCommand.NONE,
+    var state: IctrlState = IctrlState.NONE,
+    var errors: MutableList<IctrlError> = mutableListOf(),
+
+    var workMode: IctrlWorkMode = IctrlWorkMode.PROD,
+    var stubCase: IctrlStubs = IctrlStubs.NONE,
+
+    var requestId: IctrlRequestId = IctrlRequestId.NONE,
+    var timeStart: Instant = Instant.NONE,
+    var commodityRequest: IctrlCommodity = IctrlCommodity(),
+    var commodityFilterRequest: IctrlCommodityFilter = IctrlCommodityFilter(),
+    var commodityResponse: IctrlCommodity = IctrlCommodity(),
+    var commoditiesResponse: MutableList<IctrlCommodity> = mutableListOf(),
+)
