+++ common/src/main/kotlin/ru/otus/otuskotlin/common/models/IctrlCommodityFilter.kt
@@ -1,0 +1,6 @@
+package ru.otus.otuskotlin.common.models
+
+data class IctrlCommodityFilter(
+    var searchString: String = "",
+    var ownerId: IctrlUserId = IctrlUserId.NONE,
+)
