+++ common/src/main/kotlin/ru/otus/otuskotlin/common/models/IctrlCommodityType.kt
@@ -1,0 +1,8 @@
+package ru.otus.otuskotlin.common.models
+
+enum class IctrlCommodityType {
+    TUBE_LINE_PART,
+    FASTENER_PART,
+    COATING_MATERIAL,
+    NONE,
+}
