+++ common/src/main/kotlin/ru/otus/otuskotlin/common/models/IctrlRequestId.kt
@@ -1,0 +1,10 @@
+package ru.otus.otuskotlin.common.models
+
+@JvmInline
+value class IctrlRequestId(private val id: String) {
+    fun asString() = id
+
+    companion object {
+        val NONE = IctrlRequestId("")
+    }
+}
