+++ common/src/main/kotlin/ru/otus/otuskotlin/common/models/IctrlVisibility.kt
@@ -1,0 +1,8 @@
+package ru.otus.otuskotlin.common.models
+
+enum class IctrlVisibility {
+    NONE,
+    VISIBLE_TO_OWNER,
+    VISIBLE_TO_GROUP,
+    VISIBLE_PUBLIC,
+}
