+++ common/src/main/kotlin/ru/otus/otuskotlin/common/models/IctrlWorkMode.kt
@@ -1,0 +1,7 @@
+package ru.otus.otuskotlin.common.models
+
+enum class IctrlWorkMode {
+    PROD,
+    TEST,
+    STUB,
+}
