+++ common/src/main/kotlin/ru/otus/otuskotlin/common/stubs/IctrlStubs.kt
@@ -1,0 +1,13 @@
+package ru.otus.otuskotlin.common.stubs
+
+enum class IctrlStubs {
+    NONE,
+    SUCCESS,
+    NOT_FOUND,
+    BAD_ID,
+    BAD_NAME,
+    BAD_DESCRIPTION,
+    BAD_VISIBILITY,
+    CANNOT_DELETE,
+    BAD_SEARCH_STRING,
+}
