--- common/src/main/kotlin/stubs/IctrlStubs.kt
@@ -1,13 +1,0 @@
-package stubs
-
-enum class IctrlStubs {
-    NONE,
-    SUCCESS,
-    NOT_FOUND,
-    BAD_ID,
-    BAD_NAME,
-    BAD_DESCRIPTION,
-    BAD_VISIBILITY,
-    CANNOT_DELETE,
-    BAD_SEARCH_STRING,
-}
