--- mappers-v1/src/main/kotlin/exceptions/UnknownIctrlCommand.kt
@@ -1,5 +1,0 @@
-package exceptions
-
-import models.IctrlCommand
-
-class UnknownIctrlCommand(command: IctrlCommand) : Throwable("Wrong command $command at mapping toTransport stage")
