--- mappers-v1/src/main/kotlin/exceptions/UnknownRequestClass.kt
@@ -1,3 +1,0 @@
-package exceptions
-
-class UnknownRequestClass(clazz: Class<*>) : RuntimeException("Class $clazz cannot be mapped to IctrlContext")
