--- stubs/src/main/kotlin/IctrlCommodityStub.kt
@@ -1,6 +1,0 @@
-import IctrlCommodityStubBolts.COMMODITY_BOLT
-import models.IctrlCommodity
-
-object IctrlCommodityStub {
-    fun get(): IctrlCommodity = COMMODITY_BOLT
-}
