+++ stubs/src/main/kotlin/ru/otus/otuskotlin/stubs/IctrlCommodityStub.kt
@@ -1,0 +1,8 @@
+package ru.otus.otuskotlin.stubs
+
+import ru.otus.otuskotlin.common.models.IctrlCommodity
+import ru.otus.otuskotlin.stubs.IctrlCommodityStubBolts.COMMODITY_BOLT
+
+object IctrlCommodityStub {
+    fun get(): IctrlCommodity = COMMODITY_BOLT
+}
